# Use view for RNN gate slice extraction