@sdobber 2021-11-04
@ToucheSir  I'm happy to run some benchmarks on this change, but I'm out traveling at the moment, so it will be a few days before I can get to that.