@jeremiedb 2021-11-04
@ToucheSir I ran some test mid-size LSTMs, 2 layers, 64 neurons, 256 batch size. 

Although modest, the proposal looks to bring some performance benefits: 

**CPU**
master: 
68.950850 seconds (7.97 M allocations: 62.804 GiB, 34.83% gc time)

this PR: 
63.452862 seconds (8.97 M allocations: 59.036 GiB, 31.14% gc time)

**GPU**
master: 
12.102949 seconds (26.59 M allocations: 1.690 GiB, 3.34% gc time)

this PR:
11.820140 seconds (26.77 M allocations: 1.641 GiB, 3.08% gc time)

I'm therefore all in to adopt this change.