@ToucheSir 2021-11-06
Ok, my benchmarks are in too. The code, logs and raw data are all available at https://gist.github.com/ToucheSir/2fecbe99e5b304fed11e25e42c535cc3. Because there are so many dimensions to this, I've also created an interactive figure (partially for my own sanity) [here](https://vega.github.io/editor/#/url/vega-lite/N4IgJAzgxgFgpgWwIYgFwhgF0wBwqgegIDc4BzJAOjIEtMYBXAI0poHsDp5kTykBaADZ04JAKyUAVhDYA7EABoQOJACckCCGgDaoWRrhoQAEzjEaUQ0uJJBDQ+gDCABQCqikBDiC4UTGlBMAE8cB2U2Gll-JQAzGm9jLVRtEzMLQwBdAF8lJkjjAJBInAZ-dC8fPw82HEx2WSSUl3clR1cAEQBBEGycvQMjFQgvLWtbeyMYtlUAdwACJFljOaYkKABrDwrfMsCQsJwIqI84hMblJGG4LWzc-MLi0qNtqqUaurlzqdm1Aty19YzX4nabzRbLVYbHpZPogfQIMKqWSyAD6wVCHhsdjCABkAMoAFQAsltvDtCuiDkdoiBToJEjoQEjUZToXclg9ZCUyp4ya8QO96ud8cSPABxABK7l6GSUxiQmBQqH6COebAYqksHg1giMWFw+CItAgmGodEYTAYXk1ckwcCilCgbAQBAJ6tgcDxNFUBAATDFfEw4ABOYNwMRMADMAAYACwB4wARkTcF9Yjgsd9UDEkbEUCgkYI6hmBDgAHZfXBI76U9WYono0wAByJ2MANigbaYQcjMUjUGjUFjZbbwbbbaQxjbNZI8RmKOIEBREGElkdEGIINUyF2IFZ6CgG5AMKUmHUDW+CB0oDigjtqkKKnUV-QzLR+2PsNv98fag0g0uEZPwUG8aDvOAH2VC5nyMUxzC1GFZRAZBVE2KD93CSIaUwNg2DvGgcEKJ0ontHl5UVYCQHtJ1jEiMhClUNgZkKOkCnQVZMFgZcaAALysPcP3QaZaP0XUlBkVQym0asFETX0mwUMQ5NuEAnTsBB5Cg1jJjgBUNWuDwMOEyJbC2aYpJkuSFKU30VIADxY+J6SMOoERRQAe4FGASMXQABHBhFjqRU6lIY8lCCRyEhcmgERNOA8EMwSBVUETTPE8ydFjWS2wUNtYxUtTpki5z2PPWBEp8uFnRM3Uck8XiHFAqL0FsQQ2CgFEAGkaAAIQqsJ-MCugFRoULxKgWxGu8sI2vo-4vDQX0YTq1Rrjw0KoOgSbCgc9B8ni+1TGOZagA/view) for folks to peruse.

TL;DR: it's complicated. 

(TODO going to fill this out when I have a bit more time tomorrow)

![visualization](https://user-images.githubusercontent.com/4062143/140578326-f0873601-4d59-4ab8-9ef5-c3f96406798b.png)