@mcabbott 2021-11-06
I don't know why `@view` would slow down the backward pass, I would expect no change.

The backward is going to allocate a ton, since it'll make a full copy (mostly zero) for every time you take a slice, and another to add each pair of those with `accum`. The second you could avoid with something like https://github.com/FluxML/Zygote.jl/pull/981/files#diff-cd0210083ce3136f79bee6ebca2bcca77f41a14f11b5a7a65ea1cc54803164c3R27-R34 (as an experiment, I mean! This is not safe in isolation.) 

The first is harder. One way might be to replace individual calls to `gate(x::AbstractMatrix, h, n)` with one call that makes all of them. That could act like `eachcol` and allocate just one `similar(x)` to write everything into.