@ToucheSir 2021-11-06
```julia
x = rand(2, 3);

julia> Meta.@lower @view x[:, 2]
:($(Expr(:thunk, CodeInfo(
    @ none within `top-level scope'
1 ─      goto #3 if not true
2 ─ %2 = (view)(x, :, 2)
└──      return %2
3 ─      return false
))))

f(x) = sum(@view x[:, 1])

julia> @code_adjoint f(rand(3, 4))
Zygote.Adjoint(1: (%3, %4 :: Zygote.Context, %1, %2)
  br 3 unless true
2:
  %5 = Zygote._pullback(%4, view, %2, Main.:(:), 1)
  %6 = Base.getindex(%5, 1)
  %7 = Base.getindex(%5, 2)
  br 4 (%6, 1)
3:
  br 4 (false, 2)
4: (%8, %12 :: UInt8)
  %9 = Zygote._pullback(%4, Main.sum, %8)
  %10 = Base.getindex(%9, 1)
  %11 = Base.getindex(%9, 2)
  return %10, 1: (%1)
  %2 = @12 !== 0x01
  %3 = (@11)(%1)
  %4 = Zygote.gradindex(%3, 2)
  br 3 unless %2
  br 2
2:
  br 4 (nothing)
3:
  %5 = (@7)(%4)
  %6 = Zygote.gradindex(%5, 2)
  br 4 (%6)
4: (%7)
  %8 = Zygote.tuple(nothing, %7)
  return %8)
```
I'm not sure either, perhaps more work for the compiler? Profiling showed that the lion's share of time was spent in the generated `Pullback` function, but I thought that compilation would be cached?

> One way might be to replace individual calls to gate(x::AbstractMatrix, h, n) with one call that makes all of them.

I was thinking about that yesterday as well. This monolithic kernel approach has been successfully [done before](https://developer.nvidia.com/blog/optimizing-recurrent-neural-networks-cudnn-5/), but it would require some shuffling of the RNN cell implementations to work. Having one in-place function and corresponding rrule (+ GPU methods) for all but the matmuls in https://github.com/FluxML/Flux.jl/blob/master/src/layers/recurrent.jl#L159-L165 would be very cool.