@mcabbott 2021-11-06
If I'm reading correctly, this `gate` thing takes non-overlapping blocks along the first dimesion. So the simplest fused way might be just (untested!):
```julia
multigate(x::AbstractArray, h::Int) = ntuple(n -> gate(x,h,n), 4)

@adjoint multigate(x::AbstractArray, h::Int) =
  multigate(x, h), dy -> (vcat(dy...), nothing)
```
and then
```
  four = multigate(g, o)
  input = σ.(four[1])
  forget = σ.(four[2])
  cell = tanh.(four[3])
  output = σ.(four[4])
```
Agree there's scope for much more fusing. Like `dense_bias_act`, but more so.