@mcabbott 2021-11-07
Do the two matrix multiplications `h = σ.(Wi*x .+ Wh*h .+ b)` produce something the same size? Not e.g. matrix .+ vector I mean.

If so, then this is one way you could fuse things: https://github.com/FluxML/NNlib.jl/pull/346/files#diff-0e3febb41064ef9f892dd2b52c708e0497c548275cd543ec3f95cb947fa08b1cR6-R22