@sdobber 2021-11-09
I know I'm late to the party - sorry for that. I ran some of the benchmarks from [FluxBench](https://github.com/FluxML/FluxBench.jl/), and got about 5% faster runtimes for backward passes on my CPU. I think that's pretty nice, considering that there is also other stuff going on in the models apart from calling RNNs. Unfortunately, I don't have access to a reasonable GPU to try it there. 

| Timing [ms]   | Fwd bc/gate-view   | Bwd bc/gate-view  | Fwd master         | Bwd master        | Fwd speedup        | Bwd speedup |
| ------------- | -------------- | ------------- | -------------- | ------------- | -------------- | ------- |
| DARNN         | 25.235         | 663.732       | 26.188         | 695.946       | 3.6 %          | 4.6 %   |
| LSTNet        | 3.32           | 95.005        | 3.34           | 102,026       | 0.6 %          | 6.9 %   |
| TPA-LSTM      | 5.872          | 356.226       | 7.377          | 375,483       | 20.4 %         | 5.1 %   |